+++ searchmytraining/src/main/java/com/searchmytraining/dao/FreelancerDAO.java
@@ -6,6 +6,6 @@ public interface FreelancerDAO
 {
 	public void registerFreelancer(FreelancerEntity entity);
 	public FreelancerEntity getFreelancer(Long id);
-	public Integer getMaxUserId(String idcolumn);
+	public Long getMaxUserId(String idcolumn);
 	public FreelancerEntity getFreeLancerDetByUserId(Long userid);
 }
