+++ searchmytraining/src/main/java/com/searchmytraining/dao/InstituteDAO.java
@@ -5,6 +5,6 @@ import com.searchmytraining.entity.InstituteEntity;
 public interface InstituteDAO {
 	
 	public void updateInstituteDetails(InstituteEntity entity);
-	public Integer getMaxUserId(String idcolumn);
+	public Long getMaxUserId(String idcolumn);
 	public InstituteEntity getInstituteInfo(Long userid);
 }
