+++ searchmytraining/src/main/java/com/searchmytraining/dao/TrainerDAO.java
@@ -6,6 +6,6 @@ public interface TrainerDAO {
 	
 	public void registerTrainer(TrainerEntity entity);
 	public TrainerEntity getTrainer(Long id);
-	public Integer getMaxUserId(String idcolumn);
+	public Long getMaxUserId(String idcolumn);
 	public TrainerEntity getTrainerByUserid(Long userid);
 }
