+++ searchmytraining/src/main/java/com/searchmytraining/dao/UserDAO.java
@@ -5,7 +5,7 @@ import com.searchmytraining.entity.UserEntity;
 public interface UserDAO {
 	
 	public void addUser(UserEntity entity);
-	public Integer getMaxUserId(String idcolumn);
+	public Long getMaxUserId(String idcolumn);
 	public UserEntity getUser(long l);
 	public UserEntity getUser(String username);
 	public void updateUserUUID(String username,String uuid);
